+++ Dreamline/Dreamline/Event.swift
@@ -8,11 +8,9 @@
 
 import Foundation
 
-// @NOTE: these are updates caused by updating the model,
-// so the name should reflect that (instead of being generic 'event' etc)
-
 // @TODO: rename, the word 'event' is too commonly associated with
 // a dynamic delegate system, i'd rather it just be a piece of data
+// maybe 'update'?  'message'?
 enum Event {
     case barrierAdded(Barrier)
     case barrierDestroyed(Int) // id of barrier
@@ -21,16 +19,4 @@ enum Event {
     
     case enteredLane(Lane)
     case exitedLane(Lane)
-}
-
-// @TODO: rename
-protocol EventListener {
-    func playerHitWall()
-    func playerPassedThroughGate()
-    
-    
-    // @TODO: implement
-    //func roundStart()
-    //func roundComplete()
-    //func gameOver()
 }
