+++ Dreamline/Dreamline/Sequencer.swift
@@ -8,44 +8,23 @@
 
 import Foundation
 
-/*
-enum PatternDifficulty {
-    case EASY
-}
- */
-
 enum Gate {
     case open
     case closed
 }
 
 struct Pattern {
-    let data: [Gate]    // @TODO: maybe typedef/alias this
+    let data: [Gate]
 }
-/*
- @TODO: move into authored sequencer
-struct PatternSource {
-    let patterns: [Int: [Pattern]]
-}
- */
-
-/*
- @TODO: move into authored sequencer
-struct PatternOptions {
-    let groupCount: Int
-    let groupLength: Int
-    let difficulty: Int
-}
- */
 
 protocol Sequencer {
     func getNextPattern() -> Pattern
-    // @TODO: should there be a method like: anyRemaining() -> Bool ?
+    func isDone() -> Bool
 }
 
 class RandomSequencer: Sequencer {
+    
     func getNextPattern() -> Pattern {
-        // generate a random pattern
         var gates = [Gate]()
         var allClosed = true
         var allOpen = true
@@ -54,6 +33,8 @@ class RandomSequencer: Sequencer {
             if gate == .open { allClosed = allClosed && false }
             if gate == .closed { allOpen = allOpen && false }
             
+            // @CLEANUP: this code ensures that every gate has at least
+            // one opening or one closing, hard to understand
             if i == 3 && allClosed { gates.append(.open) }
             else if i == 3 && allOpen { gates.append(.closed) }
             else { gates.append(gate) }
@@ -62,48 +43,15 @@ class RandomSequencer: Sequencer {
         return Pattern(data: gates)
     }
     
-    func randomGate() -> Gate {
-        if Math.random() > 0.5 {
+    func isDone() -> Bool {
+        return false // @HARDCODED: random sequencer will never run out of patterns
+    }
+    
+    private func randomGate() -> Gate {
+        if Double.random() > 0.5 {
             return .open
         } else {
             return .closed
         }
     }
 }
-
-// @TODO: implement authored sequencer
-/*
-class AuthoredSequencer: Sequencer {
-    func generatePatterns(source: PatternSource, options: PatternOptions) -> [Pattern] {
-        var patterns = [Pattern]()
-        
-        for _ in 0...options.groupCount {
-            for _ in 0...options.groupLength {
-                patterns.append(self.searchForPattern(source: source, difficulty: options.difficulty))
-            }
-            /*
-            for _ in 0...options.groupPadding {
-                patterns.append(self.emptyPattern())
-            }
-            */
-        }
-        
-        return patterns
-    }
-    
-    // @FIXME: for now, it just returns a random pattern of the correct difficulty
-    private func searchForPattern(source: PatternSource, difficulty: Int) -> Pattern {
-        assert(source.patterns[difficulty] != nil)
-        
-        let patterns = source.patterns[difficulty]!
-        return patterns[Math.randInt(min: 0, max: patterns.count-1)]
-    }
-    
-    private func emptyPattern() -> Pattern {
-        return Pattern(data: [.closed, .closed, .closed])
-    }
-}
- */
-
-
-
